feat: project structure update
