+++ k8s/log/kube-logger-go/README.md
@@ -1,0 +1,18 @@
+<h2 align="center">
+    <a href="https://httpie.io" target="blank_">
+        <img height="100" alt="nullplatform" src="https://nullplatform.com/favicon/android-chrome-192x192.png" />
+    </a>
+    <br>
+    Logs Extraction for Kubernetes
+    <br>
+</h2>
+
+
+## How do I built it?
+
+After any code changes, you need to rebuild the binary.
+
+1.  Run the makefile to build the binary:
+    ```bash
+    make build
+    ```
