+++ k8s/log/kube-logger-go/go.mod
@@ -1,4 +1,4 @@
-module kube-logger
+module kube-logger-go
 
 go 1.21
