+++ k8s/log/log
@@ -3,7 +3,7 @@
 PLATFORM=$(uname | tr '[:upper:]' '[:lower:]')
 ARCH=$(uname -m)
 
-KUBE_LOGGER_SCRIPT="$SERVICE_PATH/log/kube-logger-go/bin/$PLATFORM/log-$ARCH"
+KUBE_LOGGER_SCRIPT="$SERVICE_PATH/log/kube-logger-go/bin/$PLATFORM/exec-$ARCH"
 
 if [ ! -f "$KUBE_LOGGER_SCRIPT" ]; then
     echo "Error: kube-logger bash script not found at $KUBE_LOGGER_SCRIPT" >&2
